Renaming FileReader and using it in DelimitedProtoStreamReader

+++ disk-buffering/src/main/java/io/opentelemetry/contrib/disk/buffering/internal/storage/files/reader/ReadResult.java
@@ -13,9 +13,9 @@ public final class ReadResult {
    * The total amount of data read from the stream. This number can be greater than the content
    * length as it also takes into account any delimiters size.
    */
-  public final int totalReadLength;
+  public final long totalReadLength;
 
-  public ReadResult(byte[] content, int totalReadLength) {
+  public ReadResult(byte[] content, long totalReadLength) {
     this.content = content;
     this.totalReadLength = totalReadLength;
   }
